use of 64 bits timers at rate limiter

+++ code/declarations.hpp
@@ -1239,7 +1239,7 @@ struct leakyBucket_s
 {
 	netadrtype_t type;
 	unsigned char adr[4];
-	int lastTime;
+	uint64_t lastTime;
 	signed char	burst;
 	long hash;
 	leakyBucket_t *prev, *next;
